+++ R/strat_to_time.list.R
@@ -13,11 +13,11 @@ strat_to_time.list = function(obj, x, ...){
   #' 
   #' @param obj a list with one element named "h", which will be interpreted as stratigraphic positions
   #' @param x an _adm_ object
-  #' @param ... options passed to _get_time_
+  #' @param ... options passed to `get_time`
   #' 
-  #' @seealso [time_to_strat.list()] for the transformation from time to height domain
+  #' @seealso [time_to_strat.list()] for the transformation from time to height domain, [get_time()] for the underlying procedure, [time_to_strat()] for the higher level function
   #' 
-  #' @returns a list with one named element "t" instead of the element "h". This element contains the times of the stratigraphic positions in "h".
+  #' @returns a `timelist` (inherits from `list`). A list with one named element "t" instead of the element "h". This element contains the times of the stratigraphic positions in "h".
   #' 
   #' @examples
     #' # see vignette("admtools") for an example
@@ -35,5 +35,6 @@ strat_to_time.list = function(obj, x, ...){
   
   li[["t"]]  = get_time(adm, h = li[["h"]], ...)
   li = li[names(li) != "h" ]
+  class(li) = c("timelist", "list")
   return(li)
 }